inline export default function; move employers to data/
