+++ src/components/AcheivementsContainer.tsx
@@ -1,7 +1,7 @@
 import ExperienceContainer from "./experience/ExperienceContainer";
 import ProjectsContainer from "./projects/ProjectsContainer";
 
-function AcheivementsContainer() {
+export default function AcheivementsContainer() {
 	return (
 		<div className="flex w-full flex-col gap-12">
 			<ExperienceContainer />
@@ -9,5 +9,3 @@ function AcheivementsContainer() {
 		</div>
 	);
 }
-
-export default AcheivementsContainer;
