+++ src/data/experience.ts
@@ -1,4 +1,4 @@
-import employers from "@components/Employers";
+import employers from "@data/employers";
 
 export type CareerEvent = {
 	event: string;
